gpio: only unexport in close if exported in open in sysfs implementation
During open, the current implementation tries to export the GPIO if it
is not already available, but then unconditionally unexports the GPIO
again during close. This causes issues if exporting and unexporting is
controlled by some external mechanism. This was no problem in v1.x,
since the unexport code was only added later.

resolves #39.

Signed-off-by: Vanya A. Sergeev <v@sergeev.io>